src/tools/requirements.txt
PyQt6
pyyaml 